requirements.txt
Flask
openai
google-generativeai
pydantic
gtts
ffmpeg-python
python-pptx
dotenv


